+++ tests/m269_tests.sh
@@ -2,7 +2,7 @@
 
 # this script is meant to be executed from the project's root directory
 folder="$HOME/GitHub/M269-NE/book-25j"
-cmd="python allowed/allowed.py -f --file-unit (\d\d) -c m269-25j"
+cmd="python -m allowed.allowed -f --file-unit (\d\d) -c m269-25j"
 
 if [ "$1" = "run" ]; then
     echo "NOTEBOOKS"; echo