+++ tests/sample.py
@@ -7,6 +7,7 @@ This file is NOT meant to cover all Python constructs.
 import math
 import types
 from random import choice, shuffle
+from typing import Any, Iterable
 
 # Atomic types: literals, operators, functions
 
@@ -25,9 +26,12 @@ True if 1 > 0 else False
 
 # Strings
 s = "hello" + str(123)
-f : float
+f: float
 f"{f}3" in s
 s.upper()
+"abc".upper()
+s[2].join(["A", "B", "C"])
+("" + " ").count(1)
 
 # Lists
 l = [] + list("abc") + [letter for letter in "abc"]
@@ -51,8 +55,8 @@ items.discard(9)
 items.union(l)
 
 # Dictionaries
-d : dict[str, int] = {"a": 1, "b": 2}
-d["c"] : int = int(f) + int("3") # duplicate constructs are reported once per line
+d: dict[str, int] = {"a": 1, "b": 2}
+d["c"]: int = int(f) + int("3")  # duplicate constructs are reported once per line
 d.pop("a")
 for key, value in d.items():
     print(key, value)
@@ -79,6 +83,25 @@ except AssertionError as error:
 def whatever(values: list[int]) -> bool:
     """Pointless function."""
     return True
+
+
+def some_function(strings: Iterable[str]) -> list[str]:
+    list_of_str: list[str] = []
+    for string in strings:
+        list_of_str.append(string.strip().lower())  # chained methods
+    return list_of_str
+
+
+def another_test(txt: str):
+    print(txt[0].upper())
+
+def text_found(text, substring: str) -> bool:  # Ruff flags missing type hint
+    return text.find(substring) != -1  # find() isn't reported
+
+def text_found_2(text: str, substring: str) -> bool:
+    text_copy: Any = text # Any is reported in line 10
+    text_copy.find(substring) != -1  # find() isn't reported
+    return [text][0].find(substring) != -1  # find() is reported
 
 
 # Imported methods